refactor(playback): split diagnostics utilities Entire-Checkpoint: f957cd9849e0
