+++ libs/ui/playback/src/lib/playback-diagnostics/playback-diagnostics.model.ts
@@ -1,0 +1,97 @@
+import type {
+    ExternalPlayerName,
+    ResolvedPortalPlayback,
+} from 'shared-interfaces';
+
+export const PlaybackDiagnosticCode = {
+    UnsupportedContainer: 'unsupported-container',
+    UnsupportedCodec: 'unsupported-codec',
+    MediaDecodeError: 'media-decode-error',
+    NetworkError: 'network-error',
+    BrowserAccessError: 'browser-access-error',
+    DrmOrEncryption: 'drm-or-encryption',
+    UnknownPlaybackError: 'unknown-playback-error',
+} as const;
+
+export type PlaybackDiagnosticCode =
+    (typeof PlaybackDiagnosticCode)[keyof typeof PlaybackDiagnosticCode];
+
+export const PlaybackDiagnosticSource = {
+    Source: 'source',
+    Native: 'native',
+    Hls: 'hls',
+    MpegTs: 'mpegts',
+} as const;
+
+export type PlaybackDiagnosticSource =
+    (typeof PlaybackDiagnosticSource)[keyof typeof PlaybackDiagnosticSource];
+
+export const InlinePlaybackPlayer = {
+    VideoJs: 'videojs',
+    Html5: 'html5',
+    ArtPlayer: 'artplayer',
+} as const;
+
+export type InlinePlaybackPlayer =
+    (typeof InlinePlaybackPlayer)[keyof typeof InlinePlaybackPlayer];
+
+export interface PlaybackSourceMetadataInput {
+    readonly url: string;
+    readonly mimeType?: string;
+    readonly player?: InlinePlaybackPlayer;
+    readonly audioCodecs?: readonly string[];
+    readonly videoCodecs?: readonly string[];
+}
+
+export interface PlaybackSourceMetadata {
+    readonly url: string;
+    readonly extension: string;
+    readonly container: string;
+    readonly mimeType?: string;
+    readonly player?: InlinePlaybackPlayer;
+    readonly audioCodecs: readonly string[];
+    readonly videoCodecs: readonly string[];
+}
+
+export interface NativePlaybackErrorInput {
+    readonly code?: number;
+    readonly message?: string;
+}
+
+export interface HlsPlaybackErrorInput {
+    readonly type?: string;
+    readonly details?: string;
+    readonly fatal?: boolean;
+    readonly message?: string;
+    readonly error?: unknown;
+    readonly audioCodecs?: readonly string[];
+    readonly videoCodecs?: readonly string[];
+}
+
+export interface MpegTsPlaybackErrorInput {
+    readonly type?: string;
+    readonly details?: string;
+    readonly message?: string;
+    readonly info?: unknown;
+}
+
+export interface PlaybackDiagnostic {
+    readonly code: PlaybackDiagnosticCode;
+    readonly source: PlaybackDiagnosticSource;
+    readonly sourceUrl: string;
+    readonly container: string;
+    readonly mimeType?: string;
+    readonly player?: InlinePlaybackPlayer;
+    readonly audioCodecs: readonly string[];
+    readonly videoCodecs: readonly string[];
+    readonly details?: string;
+    readonly nativeErrorCode?: number;
+    readonly nativeErrorMessage?: string;
+    readonly externalFallbackRecommended: boolean;
+}
+
+export interface PlaybackFallbackRequest {
+    readonly player: ExternalPlayerName;
+    readonly playback: ResolvedPortalPlayback;
+    readonly diagnostic: PlaybackDiagnostic;
+}
